requirements.txt
scikit-learn
matplotlib
seaborn
joblib
pandas
numpy
xgboost
PyYAML
plotly

graphviz

xgboost

streamlit